+++ data/src/main/java/com/tahn/assignment/remote/api/GithubApiService.kt
@@ -4,10 +4,11 @@ import com.tahn.assignment.model.remote.GithubUserResponse
 import retrofit2.http.GET
 import retrofit2.http.Query
 
+// Documentation: https://docs.github.com/en/rest/users/users?apiVersion=2022-11-28
 interface GithubApiService {
     @GET("users")
     suspend fun getUsers(
-        @Query("per_page") perPage: Long,
+        @Query("per_page") perPage: Int,
         @Query("since") since: Int,
     ): List<GithubUserResponse>
 }
